+++ package.json
@@ -63,6 +63,7 @@
     "tailwindcss": "^3.0.2",
     "terser-webpack-plugin": "^5.2.5",
     "url": "^0.11.0",
+    "usehooks-ts": "^2.9.1",
     "web-vitals": "^2.1.4",
     "webpack-manifest-plugin": "^4.0.2",
     "workbox-webpack-plugin": "^6.4.1"
