+++ src/interfaces/search.interface.ts
